Converted to maven
