--- MANIFEST.MF
@@ -1,5 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: Provisioning.ProvisioningCore
-Class-Path: lib/tosca/json-20160212.jar
-  lib/tosca/jyaml-1.3.jar
-
